firmware: use task notification to trigger IMU query
The problem with direct triggering from ISR is that all the tasks can
be starved of processing time if the data is being transfered all the
time — e.g. when IMU update rate is very high or when the debugger is
used to pause the execution (the IMU itself is not being paused and
it looks for the firmware like IMU is constantly triggering DRDY line
even although the last measurement is not processed yet).