firmware: first revision of LED control
Here a LED is switched on by the task and switched off by the timer ISR.
Additional software triggered interrupt is used to notify the task
because the priority of the first interrupt should be the highest to
minimize LED brightness jitter.

The disadvantage of such an approach is that both the minimal LED on
time (6 μs @ 14 MHz) and the LED control cycle time (60 μs) are quite
high, mainly due to FreeRTOS task switching latencies.