firmware: toggle leds inside the ISR
This way the context switching delays are minimized. Minimal control
cycle time for a single LED is now just 8 μs, and the minimal LED on
time is decreased to 3 μs.

The downside is the increase of the interrupt latency. However, now
there are no more such time-sensitive tasks and if they appear, the
priority of the timer interrupt can be decreased. This, of course,
will introduce more significant LED on time jitter.