firmware: make ledOutputs array public
It's ok to change them and leave in inconsistent state because they are
being latched anyway.  Unfortunately, because we declare this array now
without defining its length, vscode doesn't show array content in
debugger. It can be fixed by including board.h, but it requires to move
declaration of led structs to a new file to prevent circular dependency
between board.h and led.h.