firmware: don't compile FreeRTOS for tests
Ceedling compiles FreeRTOS if it's in source paths and then we get multiple
definition error when we are trying to mock FreeRTOS functions.

Another workaround may be to use fff plugin (#include "mock_queue.h"...) but
I'm not sure it will work because other header files are including queue.h.